+++ app/controllers/preprints/provider/preprint-detail.js
@@ -25,6 +25,7 @@ const PRE_MODERATION = 'pre-moderation';
  */
 export default Controller.extend({
     i18n: service(),
+    theme: service(),
     toast: service(),
     store: service(),
 
@@ -64,6 +65,7 @@ export default Controller.extend({
         const { location: { origin } } = window;
         return [
             origin,
+            this.get('theme.id') !== 'osf' ? `preprints/${this.get('theme.id')}` : null,
             this.get('preprint.id'),
             'download',
         ].filter(part => !!part).join('/');