+++ tests/unit/controllers/preprints/provider/preprint-detail-test.js
@@ -280,25 +280,65 @@ test('submitDecision action', function (assert) {
     });
 });
 
-test('fileDownloadURL computed property', function (assert) {
+test('fileDownloadURL computed property - non-branded provider', function (assert) {
     this.inject.service('store');
     this.inject.service('theme');
+
+    this.theme.id = 'osf';
 
     const ctrl = this.subject();
 
     run(() => {
+        const provider = this.store.createRecord('preprint-provider', {
+            name: 'osf',
+            reviewsWorkflow: 'pre-moderation',
+        });
+
         const node = this.store.createRecord('node', {
             description: 'test description',
         });
 
-        const preprint = this.store.createRecord('preprint', { node });
+        const preprint = this.store.createRecord('preprint', {
+            provider,
+            node,
+        });
+
+        ctrl.setProperties({ preprint });
+        ctrl.set('preprint.id', '6gtu');
+
+        assert.strictEqual(ctrl.get('fileDownloadURL'), 'http://localhost:4201/6gtu/download');
+    });
+});
+
+test('fileDownloadURL computed property - branded provider', function(assert) {
+    this.inject.service('store');
+    this.inject.service('theme');
+
+    this.theme.id = 'engrxiv';
+
+    const ctrl = this.subject();
+
+    run(() => {
+        const provider = this.store.createRecord('preprint-provider', {
+            name: 'engrxiv',
+            reviewsWorkflow: 'pre-moderation',
+        });
+
+        const node = this.store.createRecord('node', {
+            description: 'test description',
+        });
+
+        const preprint = this.store.createRecord('preprint', {
+            provider,
+            node,
+        });
 
         ctrl.setProperties({ preprint });
         ctrl.set('preprint.id', '6gtu');
 
         const { location: { origin } } = window;
 
-        assert.strictEqual(ctrl.get('fileDownloadURL'), `${origin}/6gtu/download`);
+        assert.strictEqual(ctrl.get('fileDownloadURL'), `${origin}/preprints/engrxiv/6gtu/download`);
     });
 });